+++ cuequivariance_jax/cuequivariance_jax/equivariant_polynomial.py
