+++ cuequivariance_jax/cuequivariance_jax/segmented_polynomials/segmented_polynomial_vanilla_impl.py
