+++ cuequivariance_jax/cuequivariance_jax/segmented_polynomials/utils.py
