+++ cuequivariance_jax/tests/equivariant_polynomial_test.py
