+++ cuequivariance_jax/tests/jax_rep_array_test.py
