+++ cuequivariance_jax/tests/segmented_polynomial_test.py
