+++ cuequivariance_jax/tests/spherical_harmonics_test.py
